accept fractions for flex (#1)

+++ lib/src/_run.dart
@@ -13,7 +13,7 @@ class Run {
   final Map<RenderBox, Size> sizes = {};
 
   /// The total flex factor of all children in this run. It is updated as children are added.
-  int flex = 0;
+  double flex = 0.0;
 
   /// The height of this run. It is updated as children are added.
   double height = 0.0;
@@ -39,6 +39,6 @@ class Run {
     sizes[child] = child.size;
     width += child.size.width + (children.length > 1 ? spacing : 0);
     height = max(height, child.size.height);
-    flex += (child.parentData as StretchWrapParentData).flex ?? 0;
+    flex += (child.parentData as StretchWrapParentData).flex ?? 0.0;
   }
 }
